file filtering based on selected files implemented. further testing and  clean up required
